Split Android cloud remote service tests
